Fix another edge case
A cell on the boundary of the map would include a nil in its neighbors list.
This is obviously inappropriate, as 'nil' isn't a neighbor.  Now it just returns
the actual existing neighboring cells.

This was manifesting itself as a power plant at the boundary causing an
exception.  This commit also adds an integration test for that scenario.